feat(CSerialPortInfoBase): windows linux mac use same vid and pid format

+++ include/CSerialPort/iutils.hpp
@@ -87,7 +87,7 @@ public:
             }
             if (!*s2)
             {
-                return (cp - src);
+                return (int)(cp - src);
             }
             cp++;
         }
